backend/requirements.txt
Flask>=1.0.0
gevent
cloudant==2.13.0
deepspeech
sox
pydub
scipy
gunicorn
pika
python-dotenv
flask-sockets
